Nearly Done
What's left: in signIn.php, TODO section on validation. in library.html,
maybe make a button that says logout that just reloads the page. Not
actually in the spec, but would be nice. General testing, run through
all the things we should be able to do and make sure they all work.